refactor: better setup when already installed

+++ setup.sh
@@ -15,8 +15,13 @@ sudo snap install nvim --classic
 
 echo "--------------------------------------------------------------------------------------"
 echo "Cloning my package manager: 'packer.nvim'"
-git clone --depth 1 https://github.com/wbthomason/packer.nvim\
- ~/.local/share/nvim/site/pack/packer/start/packer.nvim
+if [[ -d ~/.local/share/nvim/site/pack/packer/start/packer.nvim ]];
+then
+    echo "packer.nvim is already installed"
+else
+    git clone --depth 1 https://github.com/wbthomason/packer.nvim\
+        ~/.local/share/nvim/site/pack/packer/start/packer.nvim
+fi
 
 echo "--------------------------------------------------------------------------------------"
 echo "Installing ripgrep (required for Telescope)"
@@ -24,15 +29,21 @@ sudo apt-get install ripgrep -y
 
 echo "--------------------------------------------------------------------------------------"
 echo "Installing nodejs (required for Copilot)"
-sudo apt-get remove nodejs -y
-sudo apt-get update
-sudo apt-get install -y ca-certificates curl gnupg
-sudo mkdir -p /etc/apt/keyrings
-curl -fsSL https://deb.nodesource.com/gpgkey/nodesource-repo.gpg.key | sudo gpg --dearmor -o /etc/apt/keyrings/nodesource.gpg
-NODE_MAJOR=20
-echo "deb [signed-by=/etc/apt/keyrings/nodesource.gpg] https://deb.nodesource.com/node_$NODE_MAJOR.x nodistro main" | sudo tee /etc/apt/sources.list.d/nodesource.list
-sudo apt-get update
-sudo apt-get install nodejs -y
+if [[ $(node -v) == v20* ]];
+then
+    echo "nodejs 20 is already installed"
+else
+    echo "nodejs 20 is not installed"
+    sudo apt-get remove nodejs -y
+    sudo apt-get update
+    sudo apt-get install -y ca-certificates curl gnupg
+    sudo mkdir -p /etc/apt/keyrings
+    curl -fsSL https://deb.nodesource.com/gpgkey/nodesource-repo.gpg.key | sudo gpg --dearmor -o /etc/apt/keyrings/nodesource.gpg
+    NODE_MAJOR=20
+    echo "deb [signed-by=/etc/apt/keyrings/nodesource.gpg] https://deb.nodesource.com/node_$NODE_MAJOR.x nodistro main" | sudo tee /etc/apt/sources.list.d/nodesource.list
+    sudo apt-get update
+    sudo apt-get install nodejs -y
+fi
 
 if [[ -f "$HOME/github_config.sh" ]];
 then
@@ -44,8 +55,21 @@ fi
 echo "--------------------------------------------------------------------------------------"
 echo "Creating symlinks for lsp configs"
 for file in $(find lsp_configs -mindepth 1 -maxdepth 1 -type f -printf '%P\n'); do
-    echo $file
-    ln -s ~/.config/nvim/lsp_configs/$file ~/.config/$file
+    if [[ -f ~/.config/$file ]];
+    then
+        if [[ $(realpath ~/.config/$file) == $(realpath ~/.config/nvim/lsp_configs/$file) ]];
+        then
+            echo "File '~/.config/$file' is already a symlink to '~/.config/nvim/lsp_configs/$file'"
+            continue
+        else
+            echo "Replacing existing file '~/.config/$file'"
+            rm ~/.config/$file
+            ln -s ~/.config/nvim/lsp_configs/$file ~/.config/$file
+        fi
+    else
+        echo "Creating symlink '~/.config/$file'"
+        ln -s ~/.config/nvim/lsp_configs/$file ~/.config/$file
+    fi
 done
 
 echo "======================================================================================"
@@ -55,4 +79,3 @@ echo "run ':PackerSync' to install neovim packages"
 echo "run ':Copilot setup' to setup GitHub Copilot"
 echo "restart neovim"
 echo "run ':MasonToolsInstall' to install all the mason tools listed in 'after/plugin/lsp.lua'"
-echo "                         after that you should have 16 mason tools installed"
